requirements.txt

beautifulsoup4==4.5.1
python-dotenv==0.6.2
click==6.7
cookies==2.2.1
Flask==0.11.1
geopy==1.11.0
googlefinance==0.7
httplib2==0.10.3
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
oauth2==1.9.0.post1
python-forecastio==1.3.5
requests==2.0.0
responses==0.5.1
six==1.10.0
Werkzeug==0.11.11
yelp==1.0.2

